backend/requirements.txt
fastapi
uvicorn
sqlmodel
requests
selenium
google-genai
python-dotenv
yfinance
apscheduler